Time: 0 ms (100.00%) | Memory: 7.9 MB (90.00%) - LeetSync

+++ 62-unique-paths/unique-paths.cpp
@@ -11,6 +11,6 @@ public:
             results = (results* ((totalSteps-k+i)))/i;
         }
 
-        return results;
+        return static_cast<int>(results);
     }
 };
